fix(slack): slashcommand status message dedupe
If the slash command is from private channel or direct message, we
should use the `channel_id` to send reply but if it is a public channel,
we should use the `channel_fullname` so the list of channels can be
properly deduped.

Also updated `channel_translate` with one more `uniq` function on the
output to ensure that the final list of `channel_ids` are all unique.