feat(kubernetes): handle secret files with berglas
[Berglas](https://github.com/GoogleCloudPlatform/berglas) is an utility
commonly used for handling secrets in kubernetes. It can be used as a
CLI tool, an administration controller or a library.

If `.ctx.berglas_files` is defined, there will be a step to fetch the
secrets and put them into designated files, and ensure the ownership,
mode and directory mode.

Also included in this PR, is the capability to name the kubernetes jobs
with customized prefixes instead of `honeydipper-job-`.

This is a portion from the original PR #107 that doesn't involve copying
the berglas binary and use it as entry point to handle secrets in
environment variables.