feat(kubernetes): handle secrets with berglas
[Berglas](https://github.com/GoogleCloudPlatform/berglas) is an utility
commonly used for handling secrets in kubernetes. It can be used as a
CLI tool, an administration controller or a library.

We use Berglas CLI as a wrapper to our jobs that run scripts and
commands. The Berglas command will fetch the secrets from supported
backend to parse the environment variables.

To keep the existing jobs compatible, the feature is only activated when
`.ctx.use_berglas` is set to true.

Also, if `.ctx.berglas_files` is defined, there will be a step to fetch
the secrets and put them into designated files, and ensure the
ownership, mode and directory mode.