fix: extract repo_info from api result
The web driver only parse json blob when the content type is exactly
`application/json`. Some of the APIs return with
`application/vnd.api+json`, and the content is not parsed. So we have to
manually parse the content into a object using `DipperCL`, then extract
the `repo_info`.

This should be fixed in the web driver code itself. This is just a
temporary workaround, and will be removed once the fix is implemented in
driver.